Removed duplicated logic using ActivityService & separated API logic of activity and IYOW activity
