+++ api/test/test_models_exports.py
@@ -21,6 +21,7 @@ def test_api_models_exports_expected_symbols() -> None:
         "GrantOperatorRequest",
         "ImpersonationTokenResponse",
         "IyowActivityResponse",
+        "IyowStudentSubmissionRow",
         "IyowSubmissionResponse",
         "JobControlOverviewResponse",
         "JobFailuresResponse",
@@ -34,7 +35,7 @@ def test_api_models_exports_expected_symbols() -> None:
         "RosterMemberResponse",
         "RosterUploadResponse",
         "RosterUploadStatusResponse",
-        "StudentSubmissionRow",
+        "StudentRosterIdentity",
         "SubmitIyowRequest",
         "UpdateCourseRequest",
         "UpdateIyowActivityRequest",
